+++ exonum/src/encoding/error.rs
@@ -136,7 +136,7 @@ pub enum Error {
     /// Basic error support, for custom fields.
     Basic(Cow<'static, str>),
     /// Other error for custom fields.
-    Other(Box<dyn StdError>),
+    Other(Box<dyn StdError + Send + Sync + 'static>),
 }
 
 impl fmt::Display for Error {
@@ -181,8 +181,8 @@ impl StdError for Error {
     }
 }
 
-impl From<Box<dyn StdError>> for Error {
-    fn from(t: Box<dyn StdError>) -> Self {
+impl From<Box<dyn StdError + Send + Sync + 'static>> for Error {
+    fn from(t: Box<dyn StdError + Send + Sync + 'static>) -> Self {
         Error::Other(t)
     }
 }