+++ exonum/src/events/codec.rs
@@ -14,11 +14,9 @@
 
 use byteorder::{ByteOrder, LittleEndian};
 use bytes::BytesMut;
+use failure;
 use tokio_io::codec::{Decoder, Encoder};
 
-use std::io;
-
-use super::error::other_error;
 use events::noise::{NoiseWrapper, HEADER_LENGTH as NOISE_HEADER_LENGTH};
 use messages::{MessageBuffer, RawMessage, HEADER_LENGTH};
 
@@ -41,9 +39,9 @@ impl MessagesCodec {
 
 impl Decoder for MessagesCodec {
     type Item = RawMessage;
-    type Error = io::Error;
+    type Error = failure::Error;
 
-    fn decode(&mut self, buf: &mut BytesMut) -> Result<Option<Self::Item>, io::Error> {
+    fn decode(&mut self, buf: &mut BytesMut) -> Result<Option<Self::Item>, Self::Error> {
         // Read header
         if buf.len() < HEADER_LENGTH {
             return Ok(None);
@@ -51,7 +49,6 @@ impl Decoder for MessagesCodec {
 
         let len = LittleEndian::read_u32(buf) as usize;
 
-        // To fix some weird `decode()` behavior https://github.com/carllerche/bytes/issues/104
         if buf.len() < len + NOISE_HEADER_LENGTH {
             return Ok(None);
         }
@@ -59,42 +56,60 @@ impl Decoder for MessagesCodec {
         let mut buf = self.session.decrypt_msg(len, buf)?;
 
         if buf[0] != 0 {
-            return Err(other_error("Message first byte must be set to 0"));
+            bail!("A first byte of the message must be set to 0");
         }
 
         // Check payload len
         let total_len = LittleEndian::read_u32(&buf[6..10]) as usize;
 
         if total_len as u32 > self.max_message_len {
-            return Err(other_error(format!(
+            bail!(
                 "Received message is too long: {}, maximum allowed length is {} bytes",
-                total_len, self.max_message_len,
-            )));
+                total_len,
+                self.max_message_len,
+            );
         }
 
         if total_len < HEADER_LENGTH {
-            return Err(other_error(format!(
+            bail!(
                 "Received malicious message with insufficient \
                  size in header: {}, expected header size {}",
-                total_len, HEADER_LENGTH
-            )));
+                total_len,
+                HEADER_LENGTH
+            );
         }
 
-        // Read message
-        if buf.len() >= total_len {
-            let data = buf.split_to(total_len).to_vec();
-            let raw = RawMessage::new(MessageBuffer::from_vec(data));
-            return Ok(Some(raw));
+        if total_len != buf.len() {
+            bail!(
+                "Received malicious message with wrong \
+                 total_len: {}, expected message length {}",
+                total_len,
+                buf.len()
+            );
         }
-        Ok(None)
+
+        let data = buf.split_to(total_len).to_vec();
+        let raw = RawMessage::new(MessageBuffer::from_vec(data));
+        Ok(Some(raw))
+    }
+
+    fn decode_eof(&mut self, buf: &mut BytesMut) -> Result<Option<Self::Item>, Self::Error> {
+        let message = self.decode(buf)?;
+        if message.is_none() && !buf.is_empty() {
+            trace!(
+                "Bytes remaining in buffer after receiving EOF. Remained bytes length is {}",
+                buf.len()
+            )
+        }
+        Ok(message)
     }
 }
 
 impl Encoder for MessagesCodec {
     type Item = RawMessage;
-    type Error = io::Error;
+    type Error = failure::Error;
 
-    fn encode(&mut self, msg: Self::Item, buf: &mut BytesMut) -> io::Result<()> {
+    fn encode(&mut self, msg: Self::Item, buf: &mut BytesMut) -> Result<(), Self::Error> {
         self.session.encrypt_msg(msg.as_ref(), buf)?;
         Ok(())
     }
@@ -102,25 +117,22 @@ impl Encoder for MessagesCodec {
 
 #[cfg(test)]
 mod test {
-    use super::MessagesCodec;
-
     use bytes::BytesMut;
+    use failure;
+    use tokio_io::codec::{Decoder, Encoder};
+
+    use super::MessagesCodec;
     use crypto::{gen_keypair_from_seed, Seed, SEED_LENGTH};
     use events::noise::{HandshakeParams, NoiseWrapper};
     use messages::{MessageBuffer, RawMessage};
     use node::state::SharedConnectList;
-    use tokio_io::codec::{Decoder, Encoder};
 
     #[test]
     fn decode_message_valid_header_size() {
         let data = vec![0_u8, 0, 0, 0, 0, 0, 10, 0, 0, 0];
-        let mut bytes: BytesMut = BytesMut::new();
-        let (ref mut responder, ref mut initiator) = create_encrypted_codecs();
-        let raw = RawMessage::new(MessageBuffer::from_vec(data.clone()));
-        initiator.encode(raw, &mut bytes).unwrap();
 
-        match responder.decode(&mut bytes) {
-            Ok(Some(ref r)) if r == &RawMessage::new(MessageBuffer::from_vec(data)) => {}
+        match get_decoded_message(&data) {
+            Ok(Some(ref message)) if *message.as_ref() == data[..] => {}
             _ => panic!("Wrong input"),
         };
     }
@@ -128,23 +140,65 @@ mod test {
     #[test]
     fn decode_message_small_size_in_header() {
         let data = vec![0_u8, 0, 0, 0, 0, 0, 0, 0, 0, 0];
-        let mut bytes: BytesMut = BytesMut::new();
-        let (ref mut responder, ref mut initiator) = create_encrypted_codecs();
-        let raw = RawMessage::new(MessageBuffer::from_vec(data));
-        initiator.encode(raw, &mut bytes).unwrap();
 
-        assert!(responder.decode(&mut bytes).is_err());
+        assert!(get_decoded_message(&data).is_err());
     }
 
     #[test]
     fn decode_message_zero_byte() {
         let data = vec![1u8, 0, 0, 0, 0, 0, 10, 0, 0, 0];
-        let mut bytes: BytesMut = BytesMut::new();
-        let (ref mut responder, ref mut initiator) = create_encrypted_codecs();
-        let raw = RawMessage::new(MessageBuffer::from_vec(data));
 
+        assert!(get_decoded_message(&data).is_err());
+    }
+
+    #[test]
+    #[should_panic(expected = "Received malicious message with wrong total_len")]
+    fn decode_message_small_length() {
+        let data = vec![0u8, 0, 0, 0, 0, 0, 11, 0, 0, 0, 0, 0];
+
+        get_decoded_message(&data).unwrap();
+    }
+
+    #[test]
+    #[should_panic(expected = "Received malicious message with wrong total_len")]
+    fn decode_message_big_length() {
+        let data = vec![0u8, 0, 0, 0, 0, 0, 11, 0, 0, 0];
+
+        get_decoded_message(&data).unwrap();
+    }
+
+    #[test]
+    fn decode_message_eof() {
+        let (ref mut responder, ref mut initiator) = create_encrypted_codecs();
+
+        let data = vec![0u8, 0, 0, 0, 0, 0, 10, 0, 0, 0];
+        let raw = RawMessage::new(MessageBuffer::from_vec(data.clone()));
+
+        let mut bytes: BytesMut = BytesMut::new();
+        initiator.encode(raw.clone(), &mut bytes).unwrap();
         initiator.encode(raw, &mut bytes).unwrap();
-        assert!(responder.decode(&mut bytes).is_err());
+
+        match responder.decode_eof(&mut bytes.clone()) {
+            Ok(Some(ref message)) if *message.as_ref() == data[..] => {}
+            _ => panic!("Wrong input"),
+        };
+
+        // Emulate EOF behavior.
+        bytes.truncate(1);
+        assert!(responder.decode_eof(&mut bytes).unwrap().is_none());
+
+        bytes.clear();
+        assert!(responder.decode_eof(&mut bytes).unwrap().is_none());
+    }
+
+    fn get_decoded_message(data: &[u8]) -> Result<Option<RawMessage>, failure::Error> {
+        let (ref mut responder, ref mut initiator) = create_encrypted_codecs();
+        let raw = RawMessage::new(MessageBuffer::from_vec(data.to_vec()));
+
+        let mut bytes: BytesMut = BytesMut::new();
+        initiator.encode(raw, &mut bytes).unwrap();
+
+        responder.decode(&mut bytes)
     }
 
     fn create_encrypted_codecs() -> (MessagesCodec, MessagesCodec) {